Conflicts:
	scilab/CHANGES_5.4.X
	scilab/Scilab.sln
	scilab/Scilab_f2c.sln
	scilab/configure
	scilab/modules/Makefile.in
	scilab/modules/core/includes/callinterf.h
	scilab/modules/core/src/c/callinterf.c
	scilab/modules/core/src/c/dynamic_gateways.c
	scilab/modules/graphic_export/src/c/xs2file.c
	scilab/modules/graphics/sci_gateway/c/sci_Legend.c
	scilab/modules/graphics/sci_gateway/c/sci_addcb.c
	scilab/modules/graphics/sci_gateway/c/sci_champ.c
	scilab/modules/graphics/sci_gateway/c/sci_copy.c
	scilab/modules/graphics/sci_gateway/c/sci_delete.c
	scilab/modules/graphics/sci_gateway/c/sci_drawaxis.c
	scilab/modules/graphics/sci_gateway/c/sci_drawlater.c
	scilab/modules/graphics/sci_gateway/c/sci_drawnow.c
	scilab/modules/graphics/sci_gateway/c/sci_fec.c
	scilab/modules/graphics/sci_gateway/c/sci_get.c
	scilab/modules/graphics/sci_gateway/c/sci_glue.c
	scilab/modules/graphics/sci_gateway/c/sci_grayplot.c
	scilab/modules/graphics/sci_gateway/c/sci_move.c
	scilab/modules/graphics/sci_gateway/c/sci_newaxes.c
	scilab/modules/graphics/sci_gateway/c/sci_plot2d.c
	scilab/modules/graphics/sci_gateway/c/sci_relocate_handle.c
	scilab/modules/graphics/sci_gateway/c/sci_show_window.c
	scilab/modules/graphics/sci_gateway/c/sci_swap_handles.c
	scilab/modules/graphics/sci_gateway/c/sci_unglue.c
	scilab/modules/graphics/sci_gateway/c/sci_xarc.c
	scilab/modules/graphics/sci_gateway/c/sci_xarrows.c
	scilab/modules/graphics/sci_gateway/c/sci_xfpolys.c
	scilab/modules/graphics/sci_gateway/c/sci_xpoly.c
	scilab/modules/graphics/sci_gateway/c/sci_xrect.c
	scilab/modules/graphics/sci_gateway/c/sci_xrects.c
	scilab/modules/graphics/sci_gateway/c/sci_xsegs.c
	scilab/modules/graphics/sci_gateway/c/sci_xstring.c
	scilab/modules/graphics/src/c/BuildObjects.c
	scilab/modules/graphics/src/c/SetProperty.c
	scilab/modules/graphics/src/c/getHandleProperty/get_anti_aliasing_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_alignment_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_auto_ticks_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_clip_state_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_data_mapping_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_format_n_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_hidden_axis_color_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_position_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_text_box_mode_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_tics_direction_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_title_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_x_label_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_x_ticks_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_y_label_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_y_ticks_property.c
	scilab/modules/graphics/src/c/getHandleProperty/set_z_ticks_property.c
	scilab/modules/gui/Makefile.in
	scilab/modules/gui/sci_gateway/c/sci_getcallbackobject.c
	scilab/modules/gui/sci_gateway/c/sci_uicontrol.c
	scilab/modules/gui/sci_gateway/c/sci_uimenu.c
	scilab/modules/gui/sci_gateway/c/sci_uiwait.c
	scilab/modules/gui/src/c/gui.vcxproj
	scilab/modules/gui/src/java/org/scilab/modules/gui/bridge/helpbrowser/SwingScilabHelpBrowserViewer.java
	scilab/modules/gui/src/jni/CallScilabBridge.cpp
	scilab/modules/gui/src/jni/CallScilabBridge.hxx
	scilab/modules/pvm/Makefile.in
	scilab/modules/renderer/src/cpp/figureDrawing/FigureScilabCall.cpp
	scilab/modules/scicos_blocks/src/c/scicos_blocks.vcxproj
	scilab/modules/xcos/src/java/org/scilab/modules/xcos/block/BasicBlock.java

+++ scilab/modules/graphics/src/nographics/nographics.c
@@ -178,7 +178,7 @@ sciPointObj *sciGetPointerFromHandle (long handle)
 /*--------------------------------------------------------------------------*/
 unsigned long getHandleFromStack( size_t stackPointer )
 {
-	return 	-1;
+	return 	(unsigned long)(-1);
 }
 /*--------------------------------------------------------------------------*/
 int callSetProperty(char * pObj, size_t stackPointer, int valueType, int nbRow, int nbCol, char * propertyName )
